Remove FinderTrait and move its functions into DocContext
